Fix distcheck and rpmbuild
Make sure the dummy Italian translation is included (which fixes
rpm find_lang) and add COPYING and NEWS.
The license in the rpm spec file is changed to "GPLv2+ and BSD",
which means the application as a whole is GPLv2+, but it includes
the utility code shipped with this project (that is all BSD). This
matches what's included in the about dialog.
In the future, the generator may have a set of predefined licenses
to choose from, or default to a generic proprietary EULA.

In any case, this does not change the license of this project, which
is 3-clause BSD.